Add javadoc, remove planNodeFactory, rename tests, add util method in AggregationFunctionType
