+++ pinot-core/src/main/java/com/linkedin/pinot/core/operator/query/DictionaryBasedAggregationOperator.java
@@ -32,7 +32,14 @@ import com.linkedin.pinot.core.query.aggregation.function.customobject.MinMaxRan
 import com.linkedin.pinot.core.segment.index.readers.Dictionary;
 
 /**
- * Aggregation operator that utilizes metadata for serving aggregation queries.
+ * Aggregation operator that utilizes dictionary for serving aggregation queries.
+ * The dictionary operator is selected in the plan maker, if the query if of aggregation type min, max, minmaxrange
+ * and the column has a dictionary.
+ * We don't use this operator if the segment has star tree,
+ * as the dictionary will have aggregated values for the metrics, and dimensions will have star node value
+ *
+ * For min value, we use the first value from the dictionary
+ * For max value we use the last value from dictionary
  */
 public class DictionaryBasedAggregationOperator extends BaseOperator<IntermediateResultsBlock> {
   private static final String OPERATOR_NAME = "DictionaryBasedAggregationOperator";
