+++ pinot-core/src/main/java/com/linkedin/pinot/core/plan/maker/InstancePlanMakerImplV2.java
@@ -15,15 +15,25 @@
  */
 package com.linkedin.pinot.core.plan.maker;
 
+import com.linkedin.pinot.common.request.AggregationInfo;
 import com.linkedin.pinot.common.request.BrokerRequest;
 import com.linkedin.pinot.core.data.manager.offline.SegmentDataManager;
 import com.linkedin.pinot.core.indexsegment.IndexSegment;
+import com.linkedin.pinot.core.plan.AggregationGroupByPlanNode;
+import com.linkedin.pinot.core.plan.AggregationPlanNode;
 import com.linkedin.pinot.core.plan.CombinePlanNode;
+import com.linkedin.pinot.core.plan.DictionaryBasedAggregationPlanNode;
 import com.linkedin.pinot.core.plan.GlobalPlanImplV0;
 import com.linkedin.pinot.core.plan.InstanceResponsePlanNode;
+import com.linkedin.pinot.core.plan.MetadataBasedAggregationPlanNode;
 import com.linkedin.pinot.core.plan.Plan;
 import com.linkedin.pinot.core.plan.PlanNode;
+import com.linkedin.pinot.core.plan.SelectionPlanNode;
+import com.linkedin.pinot.core.query.aggregation.function.AggregationFunctionFactory.AggregationFunctionType;
 import com.linkedin.pinot.core.query.config.QueryExecutorConfig;
+import com.linkedin.pinot.core.segment.index.ColumnMetadata;
+import com.linkedin.pinot.core.segment.index.SegmentMetadataImpl;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.ExecutorService;
@@ -76,7 +86,25 @@ public class InstancePlanMakerImplV2 implements PlanMaker {
 
   @Override
   public PlanNode makeInnerSegmentPlan(IndexSegment indexSegment, BrokerRequest brokerRequest) {
-    return PlanNodeFactory.getPlanNode(brokerRequest, indexSegment, _maxInitialResultHolderCapacity, _numAggrGroupsLimit);
+    if (brokerRequest.isSetAggregationsInfo()) {
+
+      if (brokerRequest.isSetGroupBy()) {
+        return new AggregationGroupByPlanNode(indexSegment, brokerRequest,
+            _maxInitialResultHolderCapacity, _numAggrGroupsLimit);
+      } else {
+        if (isFitForMetadataBasedPlan(brokerRequest, indexSegment)) {
+          return new MetadataBasedAggregationPlanNode(indexSegment, brokerRequest.getAggregationsInfo());
+        } else if (isFitForDictionaryBasedPlan(brokerRequest, indexSegment)) {
+          return new DictionaryBasedAggregationPlanNode(indexSegment, brokerRequest.getAggregationsInfo());
+        } else {
+          return new AggregationPlanNode(indexSegment, brokerRequest);
+        }
+      }
+    }
+    if (brokerRequest.isSetSelections()) {
+      return new SelectionPlanNode(indexSegment, brokerRequest);
+    }
+    throw new UnsupportedOperationException("The query contains no aggregation or selection.");
   }
 
   @Override
@@ -98,5 +126,94 @@ public class InstancePlanMakerImplV2 implements PlanMaker {
     return new GlobalPlanImplV0(new InstanceResponsePlanNode(combinePlanNode));
   }
 
+  /**
+   * Helper method to identify if query is fit to be be served purely based on metadata.
+   * Currently count queries without any filters, and min max queries if the segment has min max in
+   * metadata, are supported.
+   * @param brokerRequest Broker request
+   * @param indexSegment
+   * @return True if query can be served using metadata, false otherwise.
+   */
+  public static boolean isFitForMetadataBasedPlan(BrokerRequest brokerRequest, IndexSegment indexSegment) {
+    if (brokerRequest.getFilterQuery() != null || brokerRequest.isSetGroupBy()) {
+      return false;
+    }
+
+    List<AggregationInfo> aggregationsInfo = brokerRequest.getAggregationsInfo();
+    if (aggregationsInfo == null) {
+      return false;
+    }
+    for (AggregationInfo aggInfo : aggregationsInfo) {
+      if (!isMetadataBasedAggregationFunction(aggInfo, indexSegment)) {
+        return false;
+      }
+    }
+    return true;
+  }
+
+  private static boolean isMetadataBasedAggregationFunction(AggregationInfo aggregationInfo,
+      IndexSegment indexSegment) {
+    AggregationFunctionType aggFuncType = AggregationFunctionType.valueOf(aggregationInfo.getAggregationType().toUpperCase());
+    if (aggFuncType.equals(AggregationFunctionType.COUNT)) {
+      return true;
+    }
+
+    // Use minValue and maxValue from metadata, in non star tree cases
+    // minValue and maxValue is generated from dictionary on segment load, so it won't be correct in case of star tree
+    if (aggFuncType.isOfType(AggregationFunctionType.MAX, AggregationFunctionType.MIN, AggregationFunctionType.MINMAXRANGE)
+        && !indexSegment.getSegmentMetadata().hasStarTree()) {
+      String column = aggregationInfo.getAggregationParams().get("column");
+      SegmentMetadataImpl segmentMetadata = (SegmentMetadataImpl) indexSegment.getSegmentMetadata();
+      if (segmentMetadata.getColumnMetadataMap() == null) {
+        return false;
+      }
+      ColumnMetadata columnMetadata = segmentMetadata.getColumnMetadataFor(column);
+      if (columnMetadata.getMaxValue() != null && columnMetadata.getMinValue() != null) {
+        return true;
+      }
+    }
+    return false;
+  }
+
+  /**
+   * Helper method to identify if query is fit to be be served purely based on dictionary.
+   * It can be served through dictionary only for min, max, minmaxrange queries as of now,
+   * and if a dictionary is present for the column
+   * @param brokerRequest Broker request
+   * @param indexSegment
+   * @return True if query can be served using dictionary, false otherwise.
+   */
+  public static boolean isFitForDictionaryBasedPlan(BrokerRequest brokerRequest,
+      IndexSegment indexSegment) {
+    // Skipping dictionary in case of star tree. Results from dictionary won't be correct
+    // because of aggregated values in metrics, and ALL value in dimension
+    if ((brokerRequest.getFilterQuery() != null) || brokerRequest.isSetGroupBy()
+        || indexSegment.getSegmentMetadata().hasStarTree()) {
+      return false;
+    }
+    List<AggregationInfo> aggregationsInfo = brokerRequest.getAggregationsInfo();
+    if (aggregationsInfo == null) {
+      return false;
+    }
+    for (AggregationInfo aggregationInfo : aggregationsInfo) {
+      if (!isDictionaryBasedAggregationFunction(aggregationInfo, indexSegment)) {
+        return false;
+      }
+    }
+    return true;
+  }
+
+  private static boolean isDictionaryBasedAggregationFunction(AggregationInfo aggregationInfo,
+      IndexSegment indexSegment) {
+    AggregationFunctionType aggFuncType = AggregationFunctionType.valueOf(aggregationInfo.getAggregationType().toUpperCase());
+    if (aggFuncType.isOfType(AggregationFunctionType.MAX, AggregationFunctionType.MIN, AggregationFunctionType.MINMAXRANGE)) {
+      String column = aggregationInfo.getAggregationParams().get("column");
+      SegmentMetadataImpl segmentMetadata = (SegmentMetadataImpl) indexSegment.getSegmentMetadata();
+      if (segmentMetadata.getColumnMetadataMap() != null && segmentMetadata.hasDictionary(column)) {
+        return true;
+      }
+    }
+    return false;
+  }
 
 }
